refactor: domain 관련 코드 domain 패키지로 이동
