+++ src/main/kotlin/blackjack/domain/Score.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 interface Score
