+++ src/test/kotlin/blackjack/ScoreTest.kt
@@ -1,5 +1,6 @@
 package blackjack
 
+import blackjack.domain.SingleScore
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.DescribeSpec
